[#2727] refactor(core,server): refactor CatalogOperationDispatcher (#2728)
### What changes were proposed in this pull request?

Split CatalogOperationDispatcher into separate classes based on
different operations.

### Why are the changes needed?

Fix: #2727 

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

UTs
