+++ core/src/test/java/com/datastrato/gravitino/catalog/TestOperationDispatcher.java
@@ -1,0 +1,132 @@
+/*
+ * Copyright 2023 Datastrato Pvt Ltd.
+ * This software is licensed under the Apache License version 2.
+ */
+package com.datastrato.gravitino.catalog;
+
+import static com.datastrato.gravitino.TestFilesetPropertiesMetadata.TEST_FILESET_HIDDEN_KEY;
+import static com.datastrato.gravitino.connector.BasePropertiesMetadata.GRAVITINO_MANAGED_ENTITY;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.reset;
+import static org.mockito.Mockito.spy;
+
+import com.datastrato.gravitino.Catalog;
+import com.datastrato.gravitino.Config;
+import com.datastrato.gravitino.Configs;
+import com.datastrato.gravitino.EntityStore;
+import com.datastrato.gravitino.NameIdentifier;
+import com.datastrato.gravitino.StringIdentifier;
+import com.datastrato.gravitino.exceptions.IllegalNamespaceException;
+import com.datastrato.gravitino.meta.AuditInfo;
+import com.datastrato.gravitino.meta.BaseMetalake;
+import com.datastrato.gravitino.meta.SchemaVersion;
+import com.datastrato.gravitino.storage.IdGenerator;
+import com.datastrato.gravitino.storage.RandomIdGenerator;
+import com.datastrato.gravitino.storage.memory.TestMemoryEntityStore;
+import com.google.common.collect.ImmutableMap;
+import java.io.IOException;
+import java.time.Instant;
+import java.util.Map;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.function.Executable;
+
+public abstract class TestOperationDispatcher {
+
+  protected static EntityStore entityStore;
+
+  protected static final IdGenerator idGenerator = new RandomIdGenerator();
+
+  protected static final String metalake = "metalake";
+
+  protected static final String catalog = "catalog";
+
+  protected static CatalogManager catalogManager;
+
+  private static Config config;
+
+  @BeforeAll
+  public static void setUp() throws IOException {
+    config = new Config(false) {};
+    config.set(Configs.CATALOG_LOAD_ISOLATED, false);
+
+    entityStore = spy(new TestMemoryEntityStore.InMemoryEntityStore());
+    entityStore.initialize(config);
+    entityStore.setSerDe(null);
+
+    BaseMetalake metalakeEntity =
+        BaseMetalake.builder()
+            .withId(1L)
+            .withName(metalake)
+            .withAuditInfo(
+                AuditInfo.builder().withCreator("test").withCreateTime(Instant.now()).build())
+            .withVersion(SchemaVersion.V_0_1)
+            .build();
+    entityStore.put(metalakeEntity, true);
+
+    catalogManager = new CatalogManager(config, entityStore, idGenerator);
+
+    NameIdentifier ident = NameIdentifier.of(metalake, catalog);
+    Map<String, String> props = ImmutableMap.of();
+    catalogManager.createCatalog(ident, Catalog.Type.RELATIONAL, "test", "comment", props);
+  }
+
+  @AfterAll
+  public static void tearDown() throws IOException {
+    if (entityStore != null) {
+      entityStore.close();
+      entityStore = null;
+    }
+
+    if (catalogManager != null) {
+      catalogManager.close();
+      catalogManager = null;
+    }
+  }
+
+  @BeforeEach
+  public void beforeStart() throws IOException {
+    reset(entityStore);
+  }
+
+  @Test
+  public void testGetCatalogIdentifier() {
+    OperationDispatcher dispatcher = new OperationDispatcher(null, null, null) {};
+
+    NameIdentifier id1 = NameIdentifier.of("a");
+    assertThrows(IllegalNamespaceException.class, () -> dispatcher.getCatalogIdentifier(id1));
+
+    NameIdentifier id2 = NameIdentifier.of("a", "b");
+    assertEquals(dispatcher.getCatalogIdentifier(id2), NameIdentifier.of("a", "b"));
+
+    NameIdentifier id3 = NameIdentifier.of("a", "b", "c");
+    assertEquals(dispatcher.getCatalogIdentifier(id3), NameIdentifier.of("a", "b"));
+
+    NameIdentifier id4 = NameIdentifier.of("a", "b", "c", "d");
+    assertEquals(dispatcher.getCatalogIdentifier(id4), NameIdentifier.of("a", "b"));
+
+    NameIdentifier id5 = NameIdentifier.of("a", "b", "c", "d", "e");
+    assertEquals(dispatcher.getCatalogIdentifier(id5), NameIdentifier.of("a", "b"));
+  }
+
+  void testProperties(Map<String, String> expectedProps, Map<String, String> testProps) {
+    expectedProps.forEach(
+        (k, v) -> {
+          Assertions.assertEquals(v, testProps.get(k));
+        });
+    Assertions.assertFalse(testProps.containsKey(StringIdentifier.ID_KEY));
+    Assertions.assertFalse(testProps.containsKey(GRAVITINO_MANAGED_ENTITY));
+    Assertions.assertFalse(testProps.containsKey(TEST_FILESET_HIDDEN_KEY));
+  }
+
+  void testPropertyException(Executable operation, String... errorMessage) {
+    IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, operation);
+    for (String msg : errorMessage) {
+      Assertions.assertTrue(exception.getMessage().contains(msg));
+    }
+  }
+}
