+++ core/src/test/java/com/datastrato/gravitino/catalog/TestSchemaOperationDispatcher.java
@@ -1,0 +1,199 @@
+/*
+ * Copyright 2024 Datastrato Pvt Ltd.
+ * This software is licensed under the Apache License version 2.
+ */
+package com.datastrato.gravitino.catalog;
+
+import static com.datastrato.gravitino.Entity.EntityType.SCHEMA;
+import static com.datastrato.gravitino.StringIdentifier.ID_KEY;
+import static com.datastrato.gravitino.TestBasePropertiesMetadata.COMMENT_KEY;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.reset;
+
+import com.datastrato.gravitino.NameIdentifier;
+import com.datastrato.gravitino.Namespace;
+import com.datastrato.gravitino.auth.AuthConstants;
+import com.datastrato.gravitino.exceptions.NoSuchEntityException;
+import com.datastrato.gravitino.meta.AuditInfo;
+import com.datastrato.gravitino.meta.SchemaEntity;
+import com.datastrato.gravitino.rel.Schema;
+import com.datastrato.gravitino.rel.SchemaChange;
+import com.google.common.collect.ImmutableMap;
+import java.io.IOException;
+import java.time.Instant;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+public class TestSchemaOperationDispatcher extends TestOperationDispatcher {
+
+  private static SchemaOperationDispatcher dispatcher;
+
+  @BeforeAll
+  public static void initialize() throws IOException {
+    dispatcher = new SchemaOperationDispatcher(catalogManager, entityStore, idGenerator);
+  }
+
+  @Test
+  public void testCreateAndListSchemas() throws IOException {
+    Namespace ns = Namespace.of(metalake, catalog);
+
+    NameIdentifier schemaIdent = NameIdentifier.of(ns, "schema1");
+    Map<String, String> props = ImmutableMap.of("k1", "v1", "k2", "v2");
+    Schema schema = dispatcher.createSchema(schemaIdent, "comment", props);
+
+    // Check if the created Schema's field values are correct
+    Assertions.assertEquals("schema1", schema.name());
+    Assertions.assertEquals("comment", schema.comment());
+    testProperties(props, schema.properties());
+
+    // Test required table properties exception
+    Map<String, String> illegalTableProperties =
+        new HashMap<String, String>() {
+          {
+            put("k2", "v2");
+          }
+        };
+
+    testPropertyException(
+        () -> dispatcher.createSchema(schemaIdent, "comment", illegalTableProperties),
+        "Properties are required and must be set");
+
+    // Test reserved table properties exception
+    illegalTableProperties.put(COMMENT_KEY, "table comment");
+    illegalTableProperties.put(ID_KEY, "gravitino.v1.uidfdsafdsa");
+    testPropertyException(
+        () -> dispatcher.createSchema(schemaIdent, "comment", illegalTableProperties),
+        "Properties are reserved and cannot be set",
+        "comment",
+        "gravitino.identifier");
+
+    // Check if the Schema entity is stored in the EntityStore
+    SchemaEntity schemaEntity = entityStore.get(schemaIdent, SCHEMA, SchemaEntity.class);
+    Assertions.assertNotNull(schemaEntity);
+    Assertions.assertEquals("schema1", schemaEntity.name());
+    Assertions.assertNotNull(schemaEntity.id());
+
+    Optional<NameIdentifier> ident1 =
+        Arrays.stream(dispatcher.listSchemas(ns))
+            .filter(s -> s.name().equals("schema1"))
+            .findFirst();
+    Assertions.assertTrue(ident1.isPresent());
+
+    // Test when the entity store failed to put the schema entity
+    doThrow(new IOException()).when(entityStore).put(any(), anyBoolean());
+    NameIdentifier schemaIdent2 = NameIdentifier.of(ns, "schema2");
+    Schema schema2 = dispatcher.createSchema(schemaIdent2, "comment", props);
+
+    // Check if the created Schema's field values are correct
+    Assertions.assertEquals("schema2", schema2.name());
+    Assertions.assertEquals("comment", schema2.comment());
+    testProperties(props, schema2.properties());
+
+    // Check if the Schema entity is stored in the EntityStore
+    Assertions.assertFalse(entityStore.exists(schemaIdent2, SCHEMA));
+    Assertions.assertThrows(
+        NoSuchEntityException.class,
+        () -> entityStore.get(schemaIdent2, SCHEMA, SchemaEntity.class));
+
+    // Audit info is gotten from the catalog, not from the entity store
+    Assertions.assertEquals("test", schema2.auditInfo().creator());
+  }
+
+  @Test
+  public void testCreateAndAlterSchema() throws IOException {
+    NameIdentifier schemaIdent = NameIdentifier.of(metalake, catalog, "schema21");
+    Map<String, String> props = ImmutableMap.of("k1", "v1", "k2", "v2");
+    Schema schema = dispatcher.createSchema(schemaIdent, "comment", props);
+
+    // Test immutable schema properties
+    SchemaChange[] illegalChange =
+        new SchemaChange[] {SchemaChange.setProperty(COMMENT_KEY, "new comment")};
+    testPropertyException(
+        () -> dispatcher.alterSchema(schemaIdent, illegalChange),
+        "Property comment is immutable or reserved, cannot be set");
+
+    SchemaChange[] changes =
+        new SchemaChange[] {
+          SchemaChange.setProperty("k3", "v3"), SchemaChange.removeProperty("k1")
+        };
+
+    Schema alteredSchema = dispatcher.alterSchema(schemaIdent, changes);
+    Assertions.assertEquals(schema.name(), alteredSchema.name());
+    Assertions.assertEquals(schema.comment(), alteredSchema.comment());
+    Map<String, String> expectedProps = ImmutableMap.of("k2", "v2", "k3", "v3");
+    testProperties(expectedProps, alteredSchema.properties());
+    // Audit info is gotten from gravitino entity store.
+    Assertions.assertEquals(AuthConstants.ANONYMOUS_USER, alteredSchema.auditInfo().creator());
+    Assertions.assertEquals(AuthConstants.ANONYMOUS_USER, alteredSchema.auditInfo().lastModifier());
+
+    // Case 2: Test if the schema is not found in entity store
+    doThrow(new NoSuchEntityException("mock error"))
+        .when(entityStore)
+        .update(any(), any(), any(), any());
+    Schema alteredSchema1 = dispatcher.alterSchema(schemaIdent, changes);
+    Assertions.assertEquals(schema.name(), alteredSchema1.name());
+    Assertions.assertEquals(schema.comment(), alteredSchema1.comment());
+    testProperties(expectedProps, alteredSchema1.properties());
+    // Audit info is gotten from catalog, not from the entity store
+    Assertions.assertEquals("test", alteredSchema1.auditInfo().creator());
+    Assertions.assertEquals("test", alteredSchema1.auditInfo().lastModifier());
+
+    // Case 3: Test if entity store is failed to get the schema entity
+    reset(entityStore);
+    doThrow(new IOException()).when(entityStore).update(any(), any(), any(), any());
+    Schema alteredSchema2 = dispatcher.alterSchema(schemaIdent, changes);
+    Assertions.assertEquals(schema.name(), alteredSchema2.name());
+    Assertions.assertEquals(schema.comment(), alteredSchema2.comment());
+    testProperties(expectedProps, alteredSchema2.properties());
+    // Audit info is gotten from catalog, not from the entity store
+    Assertions.assertEquals("test", alteredSchema2.auditInfo().creator());
+    Assertions.assertEquals("test", alteredSchema1.auditInfo().lastModifier());
+
+    // Case 4: Test if the fetched schema entity is matched.
+    reset(entityStore);
+    SchemaEntity unmatchedEntity =
+        SchemaEntity.builder()
+            .withId(1L)
+            .withName("schema21")
+            .withNamespace(Namespace.of(metalake, catalog))
+            .withAuditInfo(
+                AuditInfo.builder()
+                    .withCreator(AuthConstants.ANONYMOUS_USER)
+                    .withCreateTime(Instant.now())
+                    .build())
+            .build();
+    doReturn(unmatchedEntity).when(entityStore).update(any(), any(), any(), any());
+    Schema alteredSchema3 = dispatcher.alterSchema(schemaIdent, changes);
+    Assertions.assertEquals(schema.name(), alteredSchema3.name());
+    Assertions.assertEquals(schema.comment(), alteredSchema3.comment());
+    testProperties(expectedProps, alteredSchema3.properties());
+    // Audit info is gotten from catalog, not from the entity store
+    Assertions.assertEquals("test", alteredSchema3.auditInfo().creator());
+    Assertions.assertEquals("test", alteredSchema1.auditInfo().lastModifier());
+  }
+
+  @Test
+  public void testCreateAndDropSchema() throws IOException {
+    NameIdentifier schemaIdent = NameIdentifier.of(metalake, catalog, "schema31");
+    Map<String, String> props = ImmutableMap.of("k1", "v1", "k2", "v2");
+
+    dispatcher.createSchema(schemaIdent, "comment", props);
+
+    boolean dropped = dispatcher.dropSchema(schemaIdent, false);
+    Assertions.assertTrue(dropped);
+
+    // Test if entity store is failed to drop the schema entity
+    dispatcher.createSchema(schemaIdent, "comment", props);
+    doThrow(new IOException()).when(entityStore).delete(any(), any(), anyBoolean());
+    Assertions.assertThrows(
+        RuntimeException.class, () -> dispatcher.dropSchema(schemaIdent, false));
+  }
+}
