+++ server/src/main/java/com/datastrato/gravitino/server/GravitinoServer.java
@@ -6,7 +6,9 @@ package com.datastrato.gravitino.server;
 
 import com.datastrato.gravitino.GravitinoEnv;
 import com.datastrato.gravitino.catalog.CatalogManager;
-import com.datastrato.gravitino.catalog.CatalogOperationDispatcher;
+import com.datastrato.gravitino.catalog.FilesetOperationDispatcher;
+import com.datastrato.gravitino.catalog.SchemaOperationDispatcher;
+import com.datastrato.gravitino.catalog.TableOperationDispatcher;
 import com.datastrato.gravitino.metalake.MetalakeManager;
 import com.datastrato.gravitino.metrics.MetricsSystem;
 import com.datastrato.gravitino.metrics.source.MetricsSource;
@@ -73,8 +75,14 @@ public class GravitinoServer extends ResourceConfig {
           protected void configure() {
             bind(gravitinoEnv.metalakesManager()).to(MetalakeManager.class).ranked(1);
             bind(gravitinoEnv.catalogManager()).to(CatalogManager.class).ranked(1);
-            bind(gravitinoEnv.catalogOperationDispatcher())
-                .to(CatalogOperationDispatcher.class)
+            bind(gravitinoEnv.schemaOperationDispatcher())
+                .to(SchemaOperationDispatcher.class)
+                .ranked(1);
+            bind(gravitinoEnv.tableOperationDispatcher())
+                .to(TableOperationDispatcher.class)
+                .ranked(1);
+            bind(gravitinoEnv.filesetOperationDispatcher())
+                .to(FilesetOperationDispatcher.class)
                 .ranked(1);
           }
         });
